# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/v1/moderation/report/{contentId}": {
      "post": {
        "operationId": "ModerationController_reportContent_v1",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
    "/api/v1/v1/moderation/appeal/{contentId}": {
      "post": {
        "operationId": "ModerationController_appealContent_v1",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppealDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Moderation"
        ]
      }
    },
...
  },
  "info": {
    "title": "Study Assistant Public API",
    "description": "Production-grade developer APIs for AI Study Assistant Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your API Key prefixed with Bearer (e.g. Bearer ska_live_...)",
        "in": "header"
      }
    },
    "schemas": {
      "ReportDto": {
        "type": "object",
        "properties": {}
      },
      "AppealDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}